Feat: DNS lookup via the proxy (DoH)
This implements a simple DNS lookup over the SOCKS5 proxy, such that we don't leak that we use Bitcoin **and** utreexo to the system-configured DNS server (usually from our ISP), or any eavesdropper.

We use DNS-over-HTTPS (DoH) which means the DNS query is encrypted. The proxy won't be able to read the contents of the query nor the response. If we use tor as proxy, then the exit node won't see the contents either.

This is also an advantage over the system DNS lookup, as most OS don't use DoT/DoH by default, they use plaintext UDP/TCP on port 53.

This simple implementation simply uses the `reqwest` crate and calls the dns.google JSON API. Google will only learn the proxy IP (or tor exit node IP) but not ours.

In `florestad` I had to change the proxy address parsing code to avoid trying to DNS-system resolve the `String` proxy if it contains the port (which defeats the whole purpose of this).