Florestad: refactor `data_dir` creation
It makes sense to have the data directory creation in its own function, especially since the `Florestad::start` method is already quite long.

Behavior is changed to parse the data directory with all '/' or '\' at the end trimmed. Also if no home directory is found we use `./.florestad`. Added tests.